% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ashRemoveGeneticPops}
\alias{ashRemoveGeneticPops}
\title{Subset built-in ash data based on populations}
\usage{
ashRemoveGeneticPops(popmap, pops)
}
\arguments{
\item{pops}{a vector of population names}
}
\description{
Subset built-in ash data based on populations
}
