% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ashSetupLandscape}
\alias{ashSetupLandscape}
\title{Setup a landscape for Ash for our simulations. everything but the surface is baked into holoSimCell}
\usage{
ashSetupLandscape(
  brickname = paste0(system.file("extdata", "rasters", package = "holoSimCell"), "/",
    "study_region_daltonIceMask_lakesMasked_linearIceSheetInterpolation.tif"),
  equalsuit = F,
  partialsuit = F,
  cellreduce = 0.45,
  xlim = NULL,
  ylim = NULL,
  timesteps = NULL
)
}
\arguments{
\item{brickname}{name of a file of a geotiff object.  layers correspond to time clicks in simulations}
}
\description{
Setup a landscape for Ash for our simulations. everything but the surface is baked into holoSimCell
}
