% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_refugia_from_abundance_raster.r
\name{assign_refugia_from_abundance_raster}
\alias{assign_refugia_from_abundance_raster}
\title{Rescale carrying capacity rasters and assign refugia}
\usage{
assign_refugia_from_abundance_raster(
  abund,
  sim,
  quant = NULL,
  threshold = NULL,
  intermediate = FALSE
)
}
\arguments{
\item{abund}{Abundance raster.}

\item{sim}{Simulation raster.}

\item{quant}{Numeric or \code{NULL}. Quantile at which to threshold the abundance raster. Values that fall above this threshold will be assumed to represent a refuge.}

\item{threshold}{Numeric or \code{NULL}. Value at which to threshold the abundance raster. Values that fall above this threshold will be assumed to represent a refuge. Note that you can either use \code{quant} or \code{threshold}, but if you specify both then \code{threshold} will be used instead (with a warning).}

\item{intermediate}{Logical. If \code{FALSE} (default), return a raster stack with refuge ID and abundance with the same extent and resolution as the simulation raster. If \code{TRUE}, return a list. One item in the list is the raster stack previously described, and the other is a stack with refuge ID and abundance at the same extent and resolution as the abundance raster.}
}
\value{
Raster stack or a list of raster stacks.
}
\description{
This function takes an "abundance" raster (i.e., from an ENM or from a pollen surface) and identifies refugia and starting (relative) abundances for each refugium. It rescales this to the extent and resolution of a "simulation" raster which typically has coarser spatial resolution than the abundance raster.
}
\details{
This function attempts to rescale the identify of refugia and abundances obtained from the abundance raster to a new spatial resolution and extent. Since the simulation raster often has cells that are larger than the cells in the abundance raster, in some cases it cannot faithfully retain abundances or even all unique refugia identified in the abundance raster. The procedure first thresholds the abundance raster using either a quantile or a threshold.  All cells that are equal to or above this threshold are assumed to be a potential refuge.  It then creates a unique "abundId" raster by assigning a unique integer number for each block of contiguous cells (using Moore neighborhood adjacency). \cr
Then, a "simAbund" raster is created with the same extent and resolution as the simulation raster. For each each cell in this raster, the function determines if it contains at least one cell in the "abundId" raster that is assigned to a refuge. The challenge here is that a single "simAbund" cell can contain cells that are assigned to multiple refugia in the "abundId" raster, and that "simAbund" cell can also include cells that have abundances that are assigned to no refugia in the abundance raster. Thus, if we simply assigned abundances to the "simAbund" cell by resampling the abundance raster, we would in some cases be too generous because a single "simAbund" cell can include cells that do not belong to this refuge. \cr
The procedure assigns abundances by first calculating a proportionality scalar where the numerator is the sum of abundances of abundance raster cells in this refugium and in the "simAbund" cell, and the denominator the sum of all abundances of all cells in this "simAbund" cell. The abundance assigned to this "simAbund" cell for this particular refugium is this scalar times the abundance from the resampling of the abundance raster to the extent/resolution of the simulation raster. Thus, abundances assigned to any particular cell in a refuge will be equal to or less than the abundance of the resampled values. \cr
The procedure then assigns each cell an integer number identifying which refugium to belongs to and an abundance corresponding to the given refuge. When cells contain more than one "abundId" refuge cell, the refuge with the greater abundance is assigned to the cell.  As a consequence, a refuge that appears in the "abundId" raster could be trimmed in extent or even eliminated if it is only represented by a few cells that have small abundances relative to a more "massive" refugium in the same cell. Also, as a result, it is possible to have distinct refugia in cells that are adjacent to one another when rescaled to the extent/resolution of the simulation raster but are spatially distinct at the scale of the abundance raster.
}
