% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_grid.R
\name{def_grid_pred}
\alias{def_grid_pred}
\title{Takes multiple spatial inputs (range shapefiles, enm rasters) and makes a grid for the genetic simulations}
\usage{
def_grid_pred(pred = NULL, samppts = NULL, init.ext = NULL,
  raster.proj = "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs",
  keep.thresh = 0.05, sim.epsg = 5070, range.epsg = 4326,
  corners = matrix(c(-1702657, 3970378, -1702657, -1e+05, 2908121,
  -1e+05, 2908212, 3970378), ncol = 2, byrow = T), base.ext = c("row",
  "col")[1])
}
\arguments{
\item{pred}{is a 3d matrix with layers of hab suitability indexed by the third dimension}

\item{samppts}{locations of actual genetic samples.  Dataframe and must have 'long' and 'lat' columns in wgs84}

\item{init.ext}{vector with the number of columns and rows hoping for in output (usually is altered in run)}

\item{raster.proj}{the proj.4 string to impose on the raster enm (if not already specified) (defaults albers)}

\item{keep.thresh}{the proportion of a cell that has to be occupied before it's included in simulation model}

\item{sim.epsg}{the epsg number that the sim runs in and upon which the resulting grid is calculated}

\item{range.epsg}{the epsg number to impose on the shapefile containing range (if not already specified) defaults wgs84)}
}
\description{
Takes multiple spatial inputs (range shapefiles, enm rasters) and makes a grid for the genetic simulations
}
