% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophist-cells2.R
\name{distancePDF}
\alias{distancePDF}
\title{simple distance density function for short and ldd}
\usage{
distancePDF(x, ssh = 1, ssc = 1, lmn = 100, lsd = 100, mix = 0)
}
\description{
returns a density for a vector of x's
}
