% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testph.R
\name{doesGmapCombine}
\alias{doesGmapCombine}
\title{Tests aggregation scheme before coalescent simulation}
\usage{
doesGmapCombine(gmap, landscape)
}
\arguments{
\item{gmap}{data frame that maps forward time populations to genetic populations (from \code{make.gmap()})}

\item{landscape}{the landscape object used in the forward simulation}
}
\value{
Returns a logical (TRUE or FALSE) indicating whether the aggregation scheme specified by gmap would combine two cells with genetic samples. Combining genetic samples can result in problems during subsequent coalescent simulations. If FALSE, no issues are expected.
}
\description{
Test the gmap aggregation scheme against a landscape to see if cells with genetic samples are being combined
}
\details{
Avoid combining cells with two separate genetic samples during cell aggregation
}
\examples{
library(holoSimCell)
parms <- drawParms(control = system.file("extdata/ashpaper","Ash_priors.csv",package="holoSimCell"))
modchoice <- 1
load(file=paste0(system.file(package="holoSimCell"),"/extdata/landscapes/",pollenPulls[[modchoice]]$file))
refpops <- pollenPulls[[modchoice]]$refs
avgCellsz <- mean(c(res(landscape$sumrast)))

ph = getpophist2.cells(h = landscape$details$ncells, xdim = landscape$details$x.dim, ydim = landscape$details$y.dim,
                       landscape=landscape,
                       refs=refpops,   
                       refsz=parms$ref_Ne,
                       lambda=parms$lambda,
                       mix=parms$mix,  
                       shortscale=parms$shortscale*avgCellsz,  
                       shortshape=parms$shortshape, 
                       longmean=parms$longmean*avgCellsz,  
                       ysz=res(landscape$sumrast)[2], 
                       xsz=res(landscape$sumrast)[1], 
                       K = parms$Ne) 

gmap=make.gmap(ph$pophist,
               xnum=2, #number of cells to aggregate in x-direction
               ynum=2) #number of aggregate in the y-direction

doesGmapCombine(gmap, landscape)

}
\seealso{
\code{\link{getpophist2.cells}}, \code{\link{make.gmap}}, \code{\link{pophist.aggregate}}, \code{\link{runFSC_step_agg3}}
}
