% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawParms.R
\name{drawParms}
\alias{drawParms}
\title{draw parameter values from a control file}
\usage{
drawParms(control = NULL)
}
\description{
takes an input file with some fixed variables and some prior distributions, outputs a list of parameter values for the simulation
}
