% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{enmScenarios}
\alias{enmScenarios}
\title{Description of alternative ENMs}
\format{
two element list (data frame and 24 element list)
}
\usage{
data(enmScenarios)
}
\description{
A two-element list that describes 24 different ENMs used in simulations from Castilla et al. (2023).
}
\details{
\itemize{
\code{enms} is a data frame with the following columns:
	\item numRefugia. The number of spatially separated refugia at the LGM
	\item gcm. The Global Circulation Model used in the ENM
	\item ext. The buffer around occurrence points used to define background regions (in km)
	\item algo. The niche modeling algorithm used in the model
	\item threshold. The threshold used to define refuge cells, cells with suitability above this value at the LGM are occupied at the start of a simulation
\item cbi. Continuous Boyce Index - a measure of model accuracy, ranges from -1 to 1, with values >0 being better than random
\item tss. True Skill Statistic - a measure of model accuracy, ranges from -1 to 1, with values >0 being better than random
\item se. Sensitivity - proportion of present sites correctly predicted when the threshold is applied
\item sp. Specificity - proportion of absent sites correctly predicted when the threshold is applied
\item auc. The area under the curve - a measure of model accuracy, ranges from 0 to 1, with values >0.5 being better than random
\item meanRefugeAbund. Average habitat suitability of cells classified as refugia (given the defined threshold)
\item rasterStackName. A text string name for the ENM, describes GCM, extent, and algorithm used. Refers to landscapes located in \code{system.file(package="holoSimCell","extdata/landscapes")}
}

\itemize{
\code{refugeCellNum} is a list with 24 elements, 1 per ENM, with each element storing a vector of IDs for refuge cells.  These ids correspond to the rasters named in 'rasterStackName' in the \code{enms} object
}
}
\keyword{datasets}
