% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holoStats_edited.R
\name{holoStats}
\alias{holoStats}
\title{Calculate summary stats for holosim}
\usage{
holoStats(out, popDF, cores = 1)
}
\arguments{
\item{out}{genetic data}

\item{popDF}{population history}

\item{extent}{size of landscape}
}
\value{
dataframe with stats
}
\description{
Calculate summary stats for holosim
}
