% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{imputed}
\alias{imputed}
\title{Genotypes for green ash (\emph{Fraxinus pennsylvanica})}
\format{
data frame with 1306 rows (loci) and 362 columns (individuals)
}
\usage{
data(imputed)
}
\description{
A set of SNP genotypes collected for green ash (\emph{Fraxinus pennsylvanica} Marshall) throughout North america.  This data frame has columns that correspond to individuals and rows that correspond to SNP loci.  The individual IDs are used in column names and mapped back to populations in \code{data(popmap)}. Genotype calls were made following alignment to the \emph{F. pennsylvanica} genome from Huff et al. 2021 - \url{https://doi.org/10.5281/zenodo.5176117}.
}
\references{
\url{https://doi.org/10.5281/zenodo.5176117}
}
\keyword{datasets}
