% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate-mig-mat.R
\name{integratedDispMat}
\alias{integratedDispMat}
\title{calc dispersal into a cell}
\usage{
integratedDispMat(xnum = 4, ynum = 4, xsz = 100, ysz = 100,
  sshp = 1, ssc = 10, mix = 0.1, nmean = 100, nvar = (nmean))
}
\description{
calculates the amount of dispersal into a cell based on the popsize and distance of other cells
}
\details{
returns a matrix with the centermost value corresponding to the source and the surrounding cells represent
the probability of dispersal into those cells.  Essentially this produces the dispersal kernel for a single
source population.  The migration matrix could be created by moving the center of this functions output to 
each population to determine dispersal into the others.
}
