% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{make.gmap}
\alias{make.gmap}
\title{make a genetic subset map for rectangular landscape}
\usage{
make.gmap(pophist, xnum = 2, ynum = 2)
}
\arguments{
\item{pophist}{a demographic population history}

\item{xnum}{number of columns to aggregate}

\item{ynum}{number of rows to aggregate}
}
\value{
a gmap object used for pophist.aggregate
}
\description{
make a genetic subset map for rectangular landscape
}
\details{
If the size of the landscape in pophist is not evenly divisible by xnum and ynum, "remainder" cells on the top and righthand sides of the landscape will be included in genetic clusters.  As a result the numbers of cells per cluster will vary
}
