% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_stats_functions.R
\name{nameStrat}
\alias{nameStrat}
\title{Rename strata in strataG object}
\usage{
nameStrat(fscout, pops, sample_pops, sample_n)
}
\description{
renames the strata to match the forward time simulation (I think)
}
