% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_grid2.R
\name{newLandscapeDim}
\alias{newLandscapeDim}
\title{resample the landscape to new dimensions given by continuous scaling number 'fac'}
\usage{
newLandscapeDim(rs, fac = 1)
}
\arguments{
\item{rs}{a raster brick object to resample}

\item{fac}{is a number that is multiplied times the number of cols to calculate new dims}
}
\description{
This function depends on enmSdm's 'rastWithSquareCells()' function to do the resampling to
a square-celled landscape of a particular resolution.  IOW the output raster has square cells.  Or at least close to square cells.
}
