% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nvec_NAs.R
\name{NvecNAs}
\alias{NvecNAs}
\title{converts cells in Nvec form to NA if their underlying hab suit is NA for that time point}
\usage{
NvecNAs(ph, l)
}
\arguments{
\item{ph}{pophist object}

\item{l}{landscape containing the rasters}
}
\description{
converts cells in Nvec form to NA if their underlying hab suit is NA for that time point
}
