% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseColHist.R
\name{parseColhist}
\alias{parseColhist}
\title{parse the colhist object that comes out of getpophist.cells into coal history}
\usage{
parseColhist(ch)
}
\arguments{
\item{ch}{colhist object}
}
\description{
parse the colhist object that comes out of getpophist.cells into coal history
}
