% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophist-cells2.R
\name{parsePopslst}
\alias{parsePopslst}
\title{parse the full demographic history to make a pophist and a coalhist object}
\usage{
parsePopslst(popslst, pops)
}
\description{
parse the full demographic history to make a pophist and a coalhist object
}
