% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellsOnSuit.R
\name{plotCellsOnSuit}
\alias{plotCellsOnSuit}
\title{hack to show the location of cells on landscape}
\usage{
plotCellsOnSuit(cells, landscape, timeslice = 1)
}
\arguments{
\item{cells}{a vector of cell ids to plot}

\item{landscape}{object containing habitat suitabiliities}
}
\description{
hack to show the location of cells on landscape
}
