% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothistslice.R
\name{plotHistSlice}
\alias{plotHistSlice}
\title{plot os slide of the landscape history}
\usage{
plotHistSlice(timeslice, ph, landscape, window = c(0, timeslice))
}
\arguments{
\item{timeslice}{this is the layer of the landscape that corresponds to a time-click}

\item{ph}{this is the population history object}

\item{landscape}{this is the landscape object that drives the simulation}
}
\description{
plot os slide of the landscape history
}
