% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{pollenPulls}
\alias{pollenPulls}
\title{Prediction surfaces for green ash pollen}
\format{
A list of 200 elements, each a list of 2 elements (character string and vector)
}
\usage{
data(pollenPulls)
}
\description{
A list of 200 pointers to files containing predictions of green ash abundance through time based on a Bayesian pollen vegetation model.  Each represents a pull from the posterior distribution of green ash abundances estimated from the model.
}
\details{
\itemize{
Each element is a list of two elements:	
	\item file. A text \code{.rda} file that stores the landscape object associated with this draw from the posterior.
	\item refs. A vector of refuge cell IDs.
}
}
\keyword{datasets}
