% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{pophist.aggregate}
\alias{pophist.aggregate}
\title{aggregate a pophist object into genetic populations}
\usage{
pophist.aggregate(ph, gmap = NULL)
}
\arguments{
\item{ph}{the output from forward time sim.  includes a pophist object}

\item{gmap}{data frame that maps forward time populations to genetic populations}
}
\value{
ph list object with new gpophist (genetic pophist) included as a list element
}
\description{
aggregate a pophist object into genetic populations
}
\details{
This function takes a fine-grained spatio-temporal history of a species generated by holoSimCells forward-time simulator and aggregates demographic populations into genetic populations that can be simulated more tractably
}
