% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{popmap}
\alias{popmap}
\title{Association between green ash IDs and populations}
\format{
data frame with 2 columns and 761 rows
}
\usage{
data(popmap)
}
\description{
This data frame provides the information to connect individual IDs (\code{id}; also found in the \code{imputed} data frame) to population IDs (\code{abbrev}) and their geographic coordinates found in \code{pts}.
}
\keyword{datasets}
