% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{pts}
\alias{pts}
\title{Geographic coordinates of green ash populations}
\format{
data frame with 5 columns and 49 rows
}
\usage{
data(pts)
}
\description{
A data frame that includes columns that link population abbreviations found in \code{popmap} to the full population names and their coordinates in WGS84 Latitude and Longitude.  Not all populations included here are found in the imputed dataset due to variability in data quality.
}
\details{
\itemize{
	\item pop. Full population name
	\item abbrev. short population name
	\item lat. Latitude in decimal degrees
	\item long. Longitude in decimal degrees
	\item N. sample size collected (this may overstate the number of individuals compared to \code{imputed} due to variable data quality.
}
}
\keyword{datasets}
