% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRep.R
\name{repParams}
\alias{repParams}
\title{specify parameters needed to run a holosim replicate}
\usage{
repParams(parms = 1)
}
\description{
Just a wrapper to specify a list of parameters that will be used in a holosim rep
}
