% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFSC_step.R
\name{runFSC_step}
\alias{runFSC_step}
\title{run fastsimcoal}
\usage{
runFSC_step(popsobj = NULL, preLGMparms = NULL, sample_pops = NULL,
  sample_n = NULL, label = "test", delete_files = TRUE,
  num_cores = 1, exec = "fsc25", loc_parms = NULL, found_Ne = NULL)
}
\description{
takes a population history and creates sequences based on the coalscent
}
