% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFSC_step_agg.R
\name{runFSC_step_agg}
\alias{runFSC_step_agg}
\title{run fastsimcoal}
\usage{
runFSC_step_agg(ph = ph, l = landscape, sample_n = NULL,
  preLGMparms = NULL, label = "test", delete_files = TRUE,
  num_cores = 1, exec = "fsc25", loc_parms = NULL, found_Ne = NULL,
  gmap = NULL)
}
\description{
takes a population history and creates sequences based on the coalscent
}
