% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFSC_NEW_agg.R
\name{runFSC_step_agg2}
\alias{runFSC_step_agg2}
\title{run fastsimcoal}
\usage{
runFSC_step_agg2(ph = ph, l = landscape, sample_n = NULL,
  preLGMparms = NULL, label = "test", delete_files = TRUE,
  num_cores = 1, exec = "fsc25", loc_parms = NULL, found_Ne = NULL,
  gmap = NULL, MAF = NULL)
}
\description{
takes a population history and creates sequences based on the coalscent
}
