% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_grid2.R
\name{transSampLoc}
\alias{transSampLoc}
\title{transform samppts to raster projection.}
\usage{
transSampLoc(samppts = NULL, range.epsg = 4326,
  raster.proj = "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs")
}
\arguments{
\item{samppts}{a dataframe with 5 cols: pop, abbrev, lat, long, and N}

\item{range.epsg}{epsg number for the projection in which the points are encoded (default 4326 -> wgs84)}

\item{raster.proj}{proj4 string for the underlying raster}
}
\description{
transform samppts to raster projection.
}
