% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseCells.R
\name{chooseRefugeCells}
\alias{chooseRefugeCells}
\title{Choose cells from a plot of suitability surface.}
\usage{
chooseRefugeCells(r)
}
\arguments{
\item{r}{this is the surface (a single raster layer)}
}
\description{
Choose cells from a plot of suitability surface.
}
