% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellsOnSuit.R
\name{plotCellsOnSuit}
\alias{plotCellsOnSuit}
\title{plot the location of cells on landscape}
\usage{
plotCellsOnSuit(cells, landscape, timeslice = 1)
}
\arguments{
\item{cells}{a vector of cell ids to plot}

\item{landscape}{the landscape object that drives the simulation (see \code{ashSetupLandscape} for a full description of the landscape object)}

\item{timeslice}{(default = 1), the layer of the landscape that corresponds to a time step (generation) in the simulation}
}
\value{
Returns a plot illustrating the simulated landscape at a particular point in time.  The simulation landscape is shown in light yellow (corresponding to open, or unglaciated, habitat) and cells queried are shown in maroon.
}
\description{
Illustration of location of specific cells on the forward simulation landscape.
}
\details{
This function allows the user to visualize the location of specific cells on the simulation landscape.  We have found this useful for confirming the location of refugial cells in our simulations.
}
\examples{
library(holoSimCell)
parms <- drawParms(control = system.file("extdata/ashpaper","Ash_priors.csv",package="holoSimCell"))
modchoice <- 1
load(file=paste0(system.file(package="holoSimCell"),"/extdata/landscapes/",pollenPulls[[modchoice]]$file))
refpops <- pollenPulls[[modchoice]]$refs
plotCellsOnSuit(refpops, landscape, timeslice = 1)

}
\seealso{
\code{\link{ashSetupLandscape}}
}
