% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothist.R
\name{plothist}
\alias{plothist}
\title{plot the population history}
\usage{
plothist(ph)
}
\arguments{
\item{ph}{population history object (direct output from gepophist.cells)}
}
\description{
plot the population history
}
