% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothist.R
\name{plothist}
\alias{plothist}
\title{plot the population history}
\usage{
plothist(ph, maxtime = NULL)
}
\arguments{
\item{ph}{population history object (direct output from gepophist.cells)}

\item{maxtime}{the maximum number of generations to plot, sets the end of the period where colonization history is plotted (i.e., maxtime=100 plots the first 100 generations of the forward simulation).}
}
\value{
Returns a multi-panel plot illustrating the history of colonization of cells on the landscape (colored arrows, red arrows are associated iwth older events, yellow and white with more recent colonization), histogram illustrating the distribution of colonization times, and a simplified plot of biotic velocity (WHAT METRIC??) through time.
}
\description{
Plots the history of colonization events from a spatially-explicit forward simulation of population expansion.
}
\details{
\itemize{
This function takes a fine-grained spatio-temporal history of a species generated by holoSimCells forward-time simulator and produces a map that shows population spread through time. The input object includes the following named components:
\item{\code{pophist}} {a data frame with population ID, cell row, cell column, timing of colonization, and source population associated with colonists.}
\item{\code{Nvecs}} {a data frame with number of generations columns and number of (aggregated) populations rows that provides population size through time from the forward demographic simulation.}
\item{\code{tmat}} {a matrix of pairwise migration rates between (aggregated) populations, used in subsequent coalescent simulations.}
\item{\code{struct}} {a vector summarizing the structure of the simulated landscape from the forward simulation, x and y dimensions of the grid, number of generations, disperal parameters, etc.}
\item{\code{hab_suit}} {a matrix with number of generations rows and number of (aggregated) populations columns that provides habitat suitability through time.}
\item{\code{coalhist}} {a data frame specifying historical events for the coalescent simulation, includes time of event (time), source (src) and sink (snk) IDs, and fractional colonization (prop).}
\item{\code{popslst}} {a list object with one element per population from the forward-time demographic simulation.  Each element includes information on location on the landscape, colonization time, source populations, and population size through time.}
}
}
\examples{
library(holoSimCell)
parms <- drawParms(control = system.file("extdata/ashpaper","Ash_priors.csv",package="holoSimCell"))
modchoice <- 1
load(file=paste0(system.file(package="holoSimCell"),"/extdata/landscapes/",pollenPulls[[modchoice]]$file))
refpops <- pollenPulls[[modchoice]]$refs
avgCellsz <- mean(c(res(landscape$sumrast)))

ph = getpophist2.cells(h = landscape$details$ncells, xdim = landscape$details$x.dim, ydim = landscape$details$y.dim,
                       hab_suit=landscape,
                       refs=refpops,   
                       refsz=parms$ref_Ne,
                       lambda=parms$lambda,
                       mix=parms$mix,  
                       shortscale=parms$shortscale*avgCellsz,  
                       shortshape=parms$shortshape, 
                       longmean=parms$longmean*avgCellsz,  
                       ysz=res(landscape$sumrast)[2], 
                       xsz=res(landscape$sumrast)[1], 
                       K = parms$Ne) 

plothist(ph)

}
\seealso{
\code{\link{getpophist2.cells}}
}
