% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothistslice.R
\name{plotHistSlice}
\alias{plotHistSlice}
\title{plot a slice of the landscape history}
\usage{
plotHistSlice(timeslice, ph, landscape, window = c(0, timeslice))
}
\arguments{
\item{timeslice}{the layer of the landscape that corresponds to a time step (generation) in the simulation}

\item{ph}{population history object (direct output from gepophist2.cells)}

\item{landscape}{the landscape object that drives the simulation (see \code{ashSetupLandscape} for a full description of the landscape object)}

\item{window}{(optional, default = c(0,timeslice)) a two-element vector that defines the window of time shown in the plot}
}
\value{
Returns a plot illustrating the history of colonization of cells on the landscape over a specified time in the simulation.  Arrows indicate the direction of colonization and points are sized relative to population size in the forward simulation.  Habitat suitability is represented by a color gradient (green as highest suitability).
}
\description{
Plots the history of colonization events from a spatially-explicit forward simulation of population expansion up to a specified time step in the simulation.
}
\details{
\itemize{
This function takes a fine-grained spatio-temporal history of a species generated by the forward-time simulator and produces a map that shows population spread through time (up to a specified time step). The input ph object includes the following named components:
\item{\code{pophist}} {a data frame with population ID, cell row, cell column, timing of colonization, and source population associated with colonists.}
\item{\code{Nvecs}} {a data frame with number of generations columns and number of (aggregated) populations rows that provides population size through time from the forward demographic simulation.}
\item{\code{tmat}} {a matrix of pairwise migration rates between (aggregated) populations, used in subsequent coalescent simulations.}
\item{\code{struct}} {a vector summarizing the structure of the simulated landscape from the forward simulation, x and y dimensions of the grid, number of generations, disperal parameters, etc.}
\item{\code{hab_suit}} {a matrix with number of generations rows and number of (aggregated) populations columns that provides habitat suitability through time.}
\item{\code{coalhist}} {a data frame specifying historical events for the coalescent simulation, includes time of event (time), source (src) and sink (snk) IDs, and fractional colonization (prop).}
\item{\code{popslst}} {a list object with one element per population from the forward-time demographic simulation.  Each element includes information on location on the landscape, colonization time, source populations, and population size through time.}
}
}
\examples{
library(holoSimCell)
parms <- drawParms(control = system.file("extdata/ashpaper","Ash_priors.csv",package="holoSimCell"))
load(file=paste0(system.file(package="holoSimCell"),"/extdata/landscapes/",pollenPulls[[1]]$file))
refpops <- pollenPulls[[1]]$refs
avgCellsz <- mean(c(res(landscape$sumrast)))

ph = getpophist2.cells(h = landscape$details$ncells, xdim = landscape$details$x.dim, ydim = landscape$details$y.dim,
                       landscape=landscape,
                       refs=refpops,   
                       refsz=parms$ref_Ne,
                       lambda=parms$lambda,
                       mix=parms$mix,  
                       shortscale=parms$shortscale*avgCellsz,  
                       shortshape=parms$shortshape, 
                       longmean=parms$longmean*avgCellsz,  
                       ysz=res(landscape$sumrast)[2], 
                       xsz=res(landscape$sumrast)[1], 
                       K = parms$Ne) 

plotHistSlice(timeslice = 20, ph = ph, landscape = landscape)
plotHistSlice(timeslice = 200, ph = ph, landscape = landscape, window = c(150,200))

}
\seealso{
\code{\link{getpophist2.cells}}, \code{\link{plothist}}, \code{\link{ashSetupLandscape}}
}
