% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFSC_step_agg3.R
\name{runFSC_step_agg3}
\alias{runFSC_step_agg3}
\title{Run fastsimcoal}
\usage{
runFSC_step_agg3(
  ph = ph,
  l = landscape,
  sample_n = NULL,
  preLGMparms = NULL,
  label = "test",
  delete_files = TRUE,
  num_cores = 1,
  exec = "fsc25",
  loc_parms = NULL,
  found_Ne = NULL,
  gmap = NULL,
  MAF = NULL,
  maxloc = 2e+05
)
}
\arguments{
\item{ph}{A pophist object, output by the forward demographic simulation function: \code{getpophist2.cells}.  Includes slots: \code{pophist}, \code{Nvecs}, \code{tmat}, \code{struct}, \code{hab_suit}, \code{coalhist}, \code{popslst}, \code{old_hab_suit}, \code{old_tmat}, \code{gmap}.  Cells in typical landscapes (>400 cells) should be aggregated after the forward-time simulation using the functions \code{make.gmap} and \code{pophist.aggregate} to improve computational efficiency.}

\item{l}{An object that defines the simulation landscape from the forward demographic simulation, includes slots: \code{details}, \code{occupied}, \code{empty}, \code{sampled}, \code{hab_suit}, \code{sumrast}, \code{samplocsrast}, \code{samplocs}, \code{sampdf}, \code{NAmask}.}

\item{sample_n}{The number of sampled (diploid) individuals per population.  Assumed to be equal across all sampled populations.}

\item{preLGMparms}{A one row data frame of parameters related to the size and history of populations in glacial refugia: \code{preLGM_t} (time in generations at which refugial populations are assumed to diverge, e.g., previous interglacial), \code{preLGM_Ne} (effective population size in the ancestral population), \code{ref_Ne} (effective size of populations in refugial cells).}

\item{label}{A text string specifying a filename prefix for output from fastsimcoal.  Note that hyphens (-) should be avoided; they are converted to periods (.) in files output by fastsimcoal.  In order to delete files with \code{strataG::fscCleanup}, avoid hyphens in file names.}

\item{delete_files}{A logical (TRUE/FALSE) indicating whether temporary simulation files and directories from fastsimcoal should be deleted after genetic data are read into R.  If TRUE, \code{strataG::fscCleanup} is used to remove files output during the coalescent simulation.}

\item{num_cores}{Number of processors to use for fastsimcoal simulations.}

\item{exec}{The filename for the fastsimcoal executable.}

\item{loc_parms}{A one row data frame of parameters providing details on the simulated genetic markers.  Contains the following named elements: \code{marker} (the type of marker to simulate, e.g., "snp"), \code{nloci} (the final number of loci included in the output dataset), \code{seq_length} (the assumed sequence length of simulated loci, if \code{marker} = "snp"), \code{mu} (the mutation rate for the simulated marker).}

\item{found_Ne}{A parameter describing the effective population size associated with newly colonized populations in the coalescent simulation.}

\item{gmap}{A map specifying a scheme for aggregating cells from the finer grained forward demographic simulation for more efficient coalescent simulation.}

\item{MAF}{A numeric value specifying a minor allele frequency filter.  SNP loci with minor alleles below this frequency are excluded from output.}

\item{maxloc}{The maximum number of loci to attempt to simulate in fastsimcoal.  If too few variable sites are generated, additional markers are simulated up to this value.}
}
\value{
Returns a data frame of individual SNP genotypes from fastsimcoal.  The first column of this data frame provides an individual ID, the second specifies the population or deme associated with the individual, and columns that follow provide diploid genotypes for simulated individuals (in two-column format).
}
\description{
Takes a population history from the forward demographic simulation and parameterizes a coalescent simulation in fastsimcoal (Excoffier et al. 2013)
}
\details{
Provides a wrapper for several helpful functions from the \code{strataG} R package.  Input parameters include a pophist object (from \code{getpophist2.cells}), a landscape object, and parameter values associated with refugial populations and genetic markers. Population size changes are limited to step-changes from founding population size (\code{found_Ne}) to the population size at the end of the forward simulation (from elements of \code{ph$Nvecs}). To improve computational efficiency, coalescent simulations are conducted for aggregated groups of cells, with an aggregation scheme defined by the function \code{make.gmap}. SNP data output by this function are subsampled to retain only one variable position from each simulated DNA segment (locus).  \emph{Note that the fastsimcoal executable must be installed and in a directory in the user's $PATH for coalescent simulations.}
}
\examples{
library(holoSimCell)
parms <- drawParms(control = system.file("extdata/ashpaper","Ash_priors.csv",package="holoSimCell"))
load(file=paste0(system.file(package="holoSimCell"),"/extdata/landscapes/",pollenPulls[[1]]$file))
refpops <- pollenPulls[[1]]$refs
avgCellsz <- mean(c(res(landscape$sumrast)))

ph = getpophist2.cells(h = landscape$details$ncells, xdim = landscape$details$x.dim, ydim = landscape$details$y.dim,
                       landscape=landscape,
                       refs=refpops,   
                       refsz=parms$ref_Ne,
                       lambda=parms$lambda,
                       mix=parms$mix,  
                       shortscale=parms$shortscale*avgCellsz,  
                       shortshape=parms$shortshape, 
                       longmean=parms$longmean*avgCellsz,  
                       ysz=res(landscape$sumrast)[2], 
                       xsz=res(landscape$sumrast)[1], 
                       K = parms$Ne) 

gmap=make.gmap(ph$pophist,
               xnum=2, #number of cells to aggregate in x-direction
               ynum=2) #number of aggregate in the y-direction

ph2 <- pophist.aggregate(ph,gmap=gmap)

loc_parms <- data.frame(marker = "snp",
                        nloci = parms$nloci,           
                        seq_length = parms$seq_length,
                        mu = parms$mu)
  
preLGMparms <- data.frame(preLGM_t = parms$preLGM_t/parms$G,   
                          preLGM_Ne = parms$preLGM_Ne,
                         ref_Ne = parms$ref_Ne)

out <- runFSC_step_agg3(ph = ph2,
                        l = landscape,
                        sample_n = 14,
                        preLGMparms = preLGMparms,
                        label = "test",
                        delete_files = TRUE,
                        num_cores = 1,
                        exec = "fsc26",
                        loc_parms = loc_parms,
                        found_Ne = parms$found_Ne,
                        gmap = gmap,
                        MAF = 0.01,
                        maxloc = 50000)

}
\seealso{
\code{\link[strataG]{fscTutorial}}, \code{\link[strataG]{fscWrite}}, \code{\link[strataG]{fscRun}}, \code{\link[strataG]{fscReadArp}}, \code{\link{getpophist2.cells}}, \code{\link{pophist.aggregate}}, \code{\link{make.gmap}}, \code{\link{holoStats}}, \url{http://cmpg.unibe.ch/software/fastsimcoal26/man/fastsimcoal26.pdf}
}
